requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
facenet-pytorch>=2.5.3
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.64.0
requests>=2.25.0
